requirements.txt
discord.py
facebook-scraper
python-dotenv
aiosqlite